backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
pydantic==2.5.1
python-dotenv==1.0.0
openai==1.3.3
google-cloud-speech==2.21.0
google-cloud-storage==2.13.0
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.0.1
